feat: Add user relationship to Product model
This commit adds a new relationship method `user()` to the `Product` model, allowing a product to belong to a user. This relationship is defined using the `belongsTo()` method and references the `User` model. This change enables easier retrieval of the user associated with a product.

Note: This commit message follows the conventional format of "type: description" and provides a clear and concise summary of the code changes.